Unpackage prepdata.
prepare_data/ is a suite of dev-only scripts. It should be in dev/,
and the python part of it should not be exposed to users, especially
since the way we'd done that was causing build failures (#195).

Instead of running

prepare_data/$SOMETHING.sh

run

dev/prepare_data/$SOMETHING.sh

And instead of running

pip install .
prepdata

run

dev/prepare_data/prepdata.py