Add a dev/.bin/ folder, and a script to add it to PATH.
This uses a Bashism, which is unfortunate. It needs $BASH_SOURCE
because $0 is set to whatever your session initially booted with
(probably /bin/sh or /bin/bash) when sourcing in scripts.
Maybe there is another option?

What this does is that now, instead of running

    python dev/prepare_data/prepdata.py --help

run

    . dev/activate

once, then any number of times

    prepdata