Removed flag -a (opacity) when calling FSLeyes
The flag was calling for a percentage (0-100), so the previous value "0.5" was not showing anything. I decided to remove that flag anyway: people can and should toggle on/off the overlay, and having the flag adds unnecessary distraction (this tutorial is on ivadomed, so we should keep non-ivadomed commands as simple as possible).