Small clarifications and typos fixes in the Unet tutorial (#610)
* Added space

* Fixed typo

* Updated FSLeyes path for consistency

* Updated "bids_path" for consistency

I didn't notice that the convention was to use the same name as the field name in the JSON config file for the example FSLeyes command

* Removed flag -a (opacity) when calling FSLeyes

The flag was calling for a percentage (0-100), so the previous value "0.5" was not showing anything. I decided to remove that flag anyway: people can and should toggle on/off the overlay, and having the flag adds unnecessary distraction (this tutorial is on ivadomed, so we should keep non-ivadomed commands as simple as possible).