`inference.py`: Fix bug when trying to access `preds`
Output is a list of two objects, but the predictions are only the first item
in the list.

I want to look deeper to see if there's a way if the second item in the list is
even needed. Can we have the model output just the first item instead? That would
remove the need for this hacky workaround.