`requirements_common.txt`: Unpin tensorboard to test
This is needed because SCT relies on TensorFlow 1.15, which
in turn relies on `tensorboard==1.15.X`, which is incompatible
with the pinned requirement.

More generally, I feel that these pinned requirements are too strict,
and that unpinning more requirements will help to avoid this problem
in the future. More thoughts in:
https://github.com/ivadomed/ivadomed/issues/501#issuecomment-719990834