Use setuptools_scm
setuptools_scm is the PyPA-blessed way to integrate git with pip.

Replaces version.txt with the most recent git tag, and MANIFEST.in with
git's contents.  This makes it much less likely to accidentally publish
a broken build, and less work to do a release.

Drops _git_info() and its subroutines. This changes the version string's
format used in dev mode (pip install -e). It previously was

  "{install_type}-{ivadomed_branch}-{ivadomed_commit}"

Now it's (roughly)

  "{tag+1}.dev{N}+g{commit}[.d{date}]"

which is https://peps.python.org/pep-0440/#version-scheme approved.

e.g. right now the most recent tag is v2.9.5, 9 commits back before the
current ec136672i, so setuptools_scm generates "2.9.6.dev9+gec136672".
If the branch is edited with uncommited changes, it is marked "dirty"
with "2.9.6.dev9+gec136672.d20220602"; this is a feature the previous
format couldn't tell us. Losing the branch name is too bad but we can
probably live without it for the benefit of conforming to standards.