Support python3.10 (#1137)
On python3.10, this requires on onnxruntime>=1.12, because that's their only version that has builds published for python 3.10. But earlier should be able to stick with earlier onnxruntimes. To support that, switch from `onnxruntime==1.7.0` to `~=1.7`, aka `>=1.7,<2`.

While we're here, also, put upper-bounds on `torch`, because we're worried about shifting too many core dependency too much all at once.

We know this dependency policy is inconsistent. We will try to clean it up in https://github.com/ivadomed/ivadomed/issues/1191.

Co-authored-by: Etienne Boucher <etienne.boucher@polymtl.ca>
Co-authored-by: Kanishk Kalra <36276423+kanishk16@users.noreply.github.com>
Co-authored-by: Yuan Wang <wang17@ualberta.ca>