feat: update default args of `wandb.login()` (#1193)
WandB includes a feature of anonymous experimental tracking wherein the user can track experiments without a WandB account. Additionally, a timeout feature exists to continue with the further execution of the code in the scenarios where an interactive prompt appears and requires input from the user, which might not be feasible in different environments, especially while running on a cluster.

These features are by default disabled but are now enabled by updating the default args of the `wandb.login()`.