`requirements.txt`: Limit `numpy<2.0` to avoid errors
This PR is meant to explicitly test Py3.8/Py3.11/Py3.12 on `master` in
its current state (i.e. prior to the `pandas` upgrade).  However,
`numpy==2.0.0` gets in the way of this, since it conflicts with pandas.

In a separate PR, I upgrade both, but for this PR, I want to keep both
at their outdated versions.